Organize relay provider
